Renamed candidate and candidateList structs for clarity
The candidate and candidateList structs are renamed distinctCandidate and distinctList for clarity. Previously, we had a point in the implementation had 2 different lists of candidate in a function. One list was represented by orderedList while the other list was represented by candidateList. The confusing part was the list represented by orderedList included all candidates for a single frame whereas the list represented by candidateList only represented the "distinct" candidates for that frame (it removed doubles). Now, the list represented by distinctList holds the "distinct" candidates