Speed of STFT in cases where it returns error greatly improved
It now will malloc its output before calling malloc for its inner variables and generating its fft plan. In one test case with interval set to 1 and paddedsize set to 10,000,000 it took over two minutes to return an error bc it first generated the fft plan for a 10,000,000 window size. Not it returns error almost immediately.