Mostly implemented allPoleGammatone
Still doesn't work well - I have more confidence in this implementation than in the naiveGammatone (carried out basic testing). I think the problem lies in the need to normalize the frequency response function. It's also possible that we need to implement a more stable way to compute the correntropy response.