Migrated existing tests from gammatoneFilter.c to check_gammatone.c
They need to be slightly refactored so that they are easier to apply to everything. One of the test_suites is presently returning a failure. I have a feeling that is due to a discrepency in the step function between the python version (that was used to determine the reference case) and the c version used in the test case.