Added some comments to ExtractMelody and ConstructNotes to explain why certain things have been commented out to allow for the use of the TransientDetectionStrategy (and what needs to be modified in the future)
Refactored TransientDetectionStrategy so that the underlying pairwise transient detection strategy can be much more easily called outside of the strategy (for use by the python bindings). I refactored part of TransientDetectionStrategy into a second function called pairwiseTransientDetection.

Added documentation for detectTransients.

Factored out the calculation of the detection function length out into a separate function called computeDetFunctionLength and modified simpleDetFunc so that the array for the detection function is pre-allocated. Updated format of simpleDetFunc's documentation.