Major update to onset detection algorithm and other fixes.
The most significant part of this commit was refactoring the correntropy
calculation in the onset detection algorithm to make better use of vector
instructions. To do this, I implemented the algorithm in terms of a very simple
vector data type and a minimal set of vector operations. I then provided 2
implementations for this type:
  1. The default choice defines the vector type as a struct and implements the
     operations calculation in terms of ordinary (scalar) C-operations. This
     choice is portable across CPU instruction sets, but is limited to
     autovectorization performed by the compiler.
  2. An alternative choice that largely just wraps SSE hardware intrincics.
     This choice will only work on x86 CPUs that support the SSSE3 (not a typo)
     instructions, but it's twice as fast as the alternative.
The user can specify their implementation preference at compile time via an
environment variable. To help test this new vector type, I introduced a new set
of tests in test/check_vector.c.

The main reason for including switchable backends is that I wrote the SSSE3
version first, and I wanted to have a record of the implementation in the
commit history. Additionally, one could imagine writing a backend that makes
use of ARM's NEON SIMD intrinsics in the future with minimal effort.
However, for simplicity, it might make more sense using just one backend.

Some of my changes to the correntropy algorithm had a cascading effect
affecting other files in the repository. One instance of this was the
improving of the const correctness of some of the transient detection helper
functions. Because of this,
- I needed to tweak the const-correctness of the gammatone filter functions
- Modifying the gammatone filter functions likewise required improving the
  const-correctness the Resample and ResampleAndAlloc functions. While doing
  this, I slightly tweaked the interface such that a pointer holding the input
  data could be directly passed to each function (rather than passing a pointer
  to a pointer). This change needed to be propagated to the gammatone filter
  functions, fVADSilenceDetection, and TransientDetectionStrategy.

Another, (more minor) instance of cascading changes occurred due to
introducing a check for the cpu's endianness (since we use an approximation for
expf in the transient detection that assumes little endianness). Since this
check is separately required for some of the unit-testing code I decided to
introduce a new utils.h header file. The idea behind this file is to hold small
inline functions that may be needed across different modules of the project.

Additonally, the way the build system worked made testing the different vector
implementations somewhat annoying. We previously used CMake to construct
"in-source" builds. Unfortunately, there's no effective way to clean up from
these builds. When I was first switching implementations, I wanted to be sure
that I was starting the build from scratch each time. This whole process was
very tedious. Therefore, I wanted to start using "out-of-source" builds. This
is the recommended approach suggested by CMake, and the intermediate build
artifacts are much easier to cleanup. This required updating some of the unit
testing code so that it could be run from anywhere (this was only relevant for
the tests that read in the input arrays and expected outputs from external).