Finished the overhaul of the cmake build system
The main focuse was on the root-level CMakeLists.txt file. I threw away the old file and rewrote it from scratch (following some basic tutorials). Only after confirming that it built the files correctly, I looked back at the old version with git and copied the code snippets to
- choose a default build mode
- set the compiler definition for the vector backend based on the environmental variable.
Note that these 2 snippets were added by myself and are not based at all on the original Check tutorial CMakelists.txt file. At this point, our CMakelists.txt file have all been rewritten and it's arguable that they are in not derived from
the CMakelists.txt file. However, because I'm not intimately familiar with the ehe precise legal definition of a derived work, I'm keeping the original license

As a result of my changes, the main CMakeLists.txt file is now MUCH shorter and easier to read. Additionally, it will now gracefully warn the users that tests can't be built when Check isn't found (instead of outright failing)

Other changes include:
- tagging the inline functions in src/utils.h, vector_scalar.h, and vector_sse.h as static functions. When we don't do this, it causes issues when compiling the library in debug mode
- fixing the include paths in tests/testtools/doubleArrayTesting.c and tests/transient/check_vector.c (they were not appropriately updated when I moved them around, and it hadn't been an issue because I had been extremely aggressive while including directories)
- tweaking the CMakeLists.txt files in the src and test directories
- tweaking the README
- adding a directory where we can collect licenses of other works that this library is derived from or other works that are directly included in this project