moved operations for writing Midi files back within extractMelody
I believe I was the person that originally separated the creation of the in-memory Midi representation and serialization into 2 distinct steps. In hindsight, this was a mistake because MIDI just specifies a serialization format.

In the future, it may be worth reconsidering this change if we arrive at a useful representation of the musical notes... (we would return that representation instead of the MIDI data)

If we really want to avoid writing the MIDI representation to disk (and just store it in memory), we can use functionallity provided by the operating system (or glibc) for creating file streams that wrap a region of memory.

In the far distant future, if we want to consider just writing MIDI representation to fixed length buffers (while reading audio data in chunks). In that case, we can just replace FILE* and fwrite with some user-defined callback